Refactored code and added new tests
