--- test/e2e/tests/account/account-custom-name.spec.ts
@@ -1,68 +1,0 @@
-import { Suite } from 'mocha';
-import { Driver } from '../../webdriver/driver';
-import { withFixtures } from '../../helpers';
-import { ACCOUNT_TYPE } from '../../constants';
-import FixtureBuilder from '../../fixture-builder';
-import { loginWithBalanceValidation } from '../../page-objects/flows/login.flow';
-import AccountDetailsModal from '../../page-objects/pages/dialog/account-details-modal';
-import AccountListPage from '../../page-objects/pages/account-list-page';
-import HeaderNavbar from '../../page-objects/pages/header-navbar';
-
-const newAccountLabel = 'Custom name';
-const anotherAccountLabel = '2nd custom name';
-
-describe('Account Custom Name Persistence', function (this: Suite) {
-  it('persists custom account label through account change and wallet lock', async function () {
-    await withFixtures(
-      {
-        fixtures: new FixtureBuilder().build(),
-        title: this.test?.fullTitle(),
-      },
-      async ({ driver }: { driver: Driver }) => {
-        await loginWithBalanceValidation(driver);
-
-        const headerNavbar = new HeaderNavbar(driver);
-        await headerNavbar.openAccountMenu();
-
-        // Change account label for existing account and verify edited account label
-        const accountListPage = new AccountListPage(driver);
-        await accountListPage.checkPageIsLoaded();
-        await accountListPage.openAccountDetailsModal('Account 1');
-
-        const accountDetailsModal = new AccountDetailsModal(driver);
-        await accountDetailsModal.checkPageIsLoaded();
-        await accountDetailsModal.changeAccountLabel(newAccountLabel);
-        await headerNavbar.checkAccountLabel(newAccountLabel);
-
-        // Add new account with custom label and verify new added account label
-        await headerNavbar.openAccountMenu();
-        await accountListPage.checkPageIsLoaded();
-        await accountListPage.addAccount({
-          accountType: ACCOUNT_TYPE.Ethereum,
-          accountName: anotherAccountLabel,
-        });
-        await headerNavbar.checkAccountLabel(anotherAccountLabel);
-
-        // Switch back to the first account and verify first custom account persists
-        await headerNavbar.openAccountMenu();
-        await accountListPage.checkPageIsLoaded();
-        await accountListPage.checkAccountDisplayedInAccountList(
-          newAccountLabel,
-        );
-        await accountListPage.switchToAccount(newAccountLabel);
-
-        // Lock and unlock wallet
-        await headerNavbar.lockMetaMask();
-        await loginWithBalanceValidation(driver);
-
-        // Verify both account labels persist after unlock
-        await headerNavbar.checkAccountLabel(newAccountLabel);
-        await headerNavbar.openAccountMenu();
-        await accountListPage.checkPageIsLoaded();
-        await accountListPage.checkAccountDisplayedInAccountList(
-          anotherAccountLabel,
-        );
-      },
-    );
-  });
-});
