+++ test/e2e/tests/multichain-accounts/account-details.spec.ts
@@ -1,28 +1,15 @@
 import { Suite } from 'mocha';
-import { WALLET_PASSWORD } from '../../helpers';
 import AccountListPage from '../../page-objects/pages/account-list-page';
 import AccountAddressModal from '../../page-objects/pages/multichain/account-address-modal';
 import AddressListModal from '../../page-objects/pages/multichain/address-list-modal';
-import PrivateKeyModal from '../../page-objects/pages/multichain/private-key-modal';
 import MultichainAccountDetailsPage from '../../page-objects/pages/multichain/multichain-account-details-page';
 import MultichainWalletDetailsPage from '../../page-objects/pages/multichain/multichain-wallet-details-page';
 import { Driver } from '../../webdriver/driver';
-import {
-  withImportedAccount,
-  withMultichainAccountsDesignEnabled,
-} from './common';
+import { withMultichainAccountsDesignEnabled } from './common';
 
 const account1 = {
   name: 'Account 1',
   address: '0x5CfE73b6021E818B776b421B1c4Db2474086a7e1',
-};
-
-const TEST_PRIVATE_KEY =
-  '14abe6f4aab7f9f626fe981c864d0adeb5685f289ac9270c27b8fd790b4235d6';
-
-const importedAccount = {
-  name: 'Imported Account 1',
-  address: '0x7A46ce51fbBB29C34aea1fE9833c27b5D2781925',
 };
 
 describe('Multichain Accounts - Account Details', function (this: Suite) {
@@ -126,70 +113,6 @@ describe('Multichain Accounts - Account Details', function (this: Suite) {
           });
 
           await accountListPage.checkAccountNameIsDisplayed(newName);
-        },
-      );
-    });
-  });
-
-  describe('View private key', function () {
-    it('shows private key when requested', async function () {
-      await withMultichainAccountsDesignEnabled(
-        {
-          title: this.test?.fullTitle(),
-        },
-        async (driver: Driver) => {
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.checkPageIsLoaded({
-            isMultichainAccountsState2Enabled: true,
-          });
-          await accountListPage.openMultichainAccountMenu({
-            accountLabel: account1.name,
-          });
-          await accountListPage.clickMultichainAccountMenuItem(
-            'Account details',
-          );
-
-          const accountDetailsPage = new MultichainAccountDetailsPage(driver);
-          await accountDetailsPage.clickPrivateKeyRow();
-
-          const privateKeyModal = new PrivateKeyModal(driver);
-          await privateKeyModal.checkPageIsLoaded();
-          await privateKeyModal.typePassword(WALLET_PASSWORD);
-          await privateKeyModal.clickConfirm();
-        },
-      );
-    });
-  });
-
-  describe('Delete private key account', function () {
-    it('removes imported private key account successfully', async function () {
-      await withImportedAccount(
-        {
-          title: this.test?.fullTitle(),
-          privateKey: TEST_PRIVATE_KEY,
-        },
-        async (driver: Driver) => {
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.checkPageIsLoaded({
-            isMultichainAccountsState2Enabled: true,
-          });
-          await accountListPage.openMultichainAccountMenu({
-            accountLabel: importedAccount.name,
-          });
-          await accountListPage.clickMultichainAccountMenuItem(
-            'Account details',
-          );
-
-          const accountDetailsPage = new MultichainAccountDetailsPage(driver);
-          await accountDetailsPage.checkPageIsLoaded();
-
-          await accountDetailsPage.clickRemoveAccountButton();
-
-          await accountDetailsPage.clickRemoveAccountConfirmButton();
-
-          await accountListPage.checkAccountIsNotDisplayedInAccountList(
-            importedAccount.name,
-          );
         },
       );
     });
