+++ test/e2e/tests/multichain-accounts/add-account.spec.ts
@@ -9,11 +9,13 @@ import LoginPage from '../../page-objects/pages/login-page';
 import ResetPasswordPage from '../../page-objects/pages/reset-password-page';
 import MultichainAccountDetailsPage from '../../page-objects/pages/multichain/multichain-account-details-page';
 import { Driver } from '../../webdriver/driver';
+import { loginWithoutBalanceValidation } from '../../page-objects/flows/login.flow';
 import { withMultichainAccountsDesignEnabled } from './common';
 
 describe('Add account', function () {
   const SECOND_ACCOUNT_NAME = 'Account 2';
   const IMPORTED_ACCOUNT_NAME = 'Imported Account 1';
+  const CUSTOM_ACCOUNT_NAME = 'Custom 1';
 
   it('should not affect public address when using secret recovery phrase to recover account with non-zero balance', async function () {
     await withMultichainAccountsDesignEnabled(
@@ -131,6 +133,49 @@ describe('Add account', function () {
         });
         await accountListPage.checkAccountNotDisplayedInAccountList(
           IMPORTED_ACCOUNT_NAME,
+        );
+      },
+    );
+  });
+
+  it('persists custom account label through account change and wallet lock', async function () {
+    await withMultichainAccountsDesignEnabled(
+      {
+        title: this.test?.fullTitle(),
+      },
+      async (driver: Driver) => {
+        const accountListPage = new AccountListPage(driver);
+        await accountListPage.checkPageIsLoaded({
+          isMultichainAccountsState2Enabled: true,
+        });
+        await accountListPage.addMultichainAccount();
+        await accountListPage.openMultichainAccountMenu({
+          accountLabel: 'Account 2',
+        });
+        await accountListPage.clickMultichainAccountMenuItem('Rename');
+        await accountListPage.changeMultichainAccountLabel(CUSTOM_ACCOUNT_NAME);
+
+        await accountListPage.checkAccountDisplayedInAccountList(
+          CUSTOM_ACCOUNT_NAME,
+        );
+        await accountListPage.switchToAccount(CUSTOM_ACCOUNT_NAME);
+
+        const headerNavbar = new HeaderNavbar(driver);
+        await headerNavbar.checkAccountLabel(CUSTOM_ACCOUNT_NAME);
+
+        // Lock and unlock wallet
+        await headerNavbar.lockMetaMask();
+        await loginWithoutBalanceValidation(driver);
+
+        // Verify both account labels persist after unlock
+        await headerNavbar.checkAccountLabel(CUSTOM_ACCOUNT_NAME);
+        await headerNavbar.openAccountsPage();
+
+        await accountListPage.checkPageIsLoaded({
+          isMultichainAccountsState2Enabled: true,
+        });
+        await accountListPage.checkAccountDisplayedInAccountList(
+          CUSTOM_ACCOUNT_NAME,
         );
       },
     );
